fix: add delete method to vacantePolicy

+++ app/Livewire/MostrarVacantes.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Vacante;
+use Illuminate\Support\Facades\Storage;
 use Livewire\Attributes\On; 
 use Livewire\Component;
 // use WithPagination;
@@ -11,6 +12,10 @@ class MostrarVacantes extends Component
 {   
     #[On('eliminarVacante')]    
     public function eliminarVacante(Vacante $vacante) {
+        // Compruebo Policy
+        $this->authorize('delete', $vacante);
+        // Elimino Imagen
+        $result = Storage::delete('public/vacantes/'.$vacante->imagen);
         $vacante->delete();
     }
